requirements.txt
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4

langchain-text-splitters>=0.3,<0.4
sentence-transformers>=2.6,<3.0
chromadb>=0.5,<0.6

pypdf>=4.0,<5.0
pymupdf>=1.23,<2.0

scikit-learn>=1.4,<2.0
numpy>=1.26,<2.0

python-dotenv>=1.0,<2.0

langchain-google-genai>=2.0,<3.0
langchain-groq>=0.2,<0.3
langchain-deepseek>=0.1,<0.2